Modified affinity flag data type in job struct.
git-svn-id: https://svn.forschooner.net/code/dnx/trunk/dnx@301 2be5450d-c085-40c6-83eb-9dc9c13f2dba

+++ common/dnxProtocol.h
@@ -205,7 +205,7 @@ typedef struct DnxNodeRequest
    DnxXID xid;                      //!< Worker node transaction ID.
    DnxReqType reqType;              //!< Request type.
    unsigned int jobCap;             //!< Job capacity.
-   double affinity;                 //!< Affinity groups bitmask.
+   unsigned long long affinity;     //!< Affinity groups bitmask.
    unsigned int ttl;                //!< Request Time-To-Live (in seconds).
    time_t expires;                  //!< Job expiration time (not transmitted).
    char address[DNX_MAX_ADDRESS];   //!< Source address.
